Rename timezone.rs to time_zone.rs
