+++ src/lib.rs
@@ -341,7 +341,7 @@ pub mod fields {
 
 pub use crate::builtins::{
     calendar::{Calendar, MonthCode},
-    core::timezone::{TimeZone, UtcOffset},
+    core::time_zone::{TimeZone, UtcOffset},
     core::DateDuration,
     Duration, Instant, PlainDate, PlainDateTime, PlainMonthDay, PlainTime, PlainYearMonth,
     ZonedDateTime,
