+++ src/options/relative_to.rs
@@ -1,7 +1,7 @@
 //! RelativeTo rounding option
 
 use crate::builtins::core::zoned_date_time::interpret_isodatetime_offset;
-use crate::builtins::core::{calendar::Calendar, timezone::TimeZone, PlainDate, ZonedDateTime};
+use crate::builtins::core::{calendar::Calendar, time_zone::TimeZone, PlainDate, ZonedDateTime};
 use crate::iso::{IsoDate, IsoTime};
 use crate::options::{Disambiguation, OffsetDisambiguation, Overflow};
 use crate::parsers::{parse_date_time, parse_zoned_date_time};
